Modificado y finalizado Ejercicio 10

+++ Tema 1/Ejemplo10.java
@@ -2,13 +2,13 @@ import java.util.Scanner;
 public class Ejemplo10 {
     public static void main (String argv[]) {
         //Declaración de variables
-        short valor;
+        double valor;
         Scanner inputValue = new Scanner(System.in);
        
         //Empieza el proceso 
 
         System.out.println("Introduce el numero");
-        valor = inputValue.nextShort();
+        valor = inputValue.nextDouble();
         inputValue.close();
         if (valor >=0){
             System.out.println("Positivo");
